+++ tuf/client/updater.py
@@ -146,6 +146,8 @@ import securesystemslib.util
 import six
 import iso8601
 
+from .handlers import ClassicMetadataHandler, FileTargetsHandler
+
 # The Timestamp role does not have signed metadata about it; otherwise we
 # would need an infinite regress of metadata. Therefore, we use some
 # default, but sane, upper file length for its metadata.
@@ -629,7 +631,9 @@ class Updater(object):
     http://www.python.org/dev/peps/pep-0008/#method-names-and-instance-variables
   """
 
-  def __init__(self, repository_name, repository_mirrors):
+  def __init__(self, repository_name, repository_mirrors,
+               metadata_handler_cls=ClassicMetadataHandler,
+               targets_handler_cls=FileTargetsHandler):
     """
     <Purpose>
       Constructor.  Instantiating an updater object causes all the metadata
@@ -719,9 +723,6 @@ class Updater(object):
     # re-downloaded.
     self.fileinfo = {}
 
-    # Store the location of the client's metadata directory.
-    self.metadata_directory = {}
-
     # Store the 'consistent_snapshot' of the Root role.  This setting
     # determines if metadata and target files downloaded from remote
     # repositories include the digest.
@@ -736,30 +737,17 @@ class Updater(object):
     # Set the path for the current set of metadata files.
     repositories_directory = tuf.settings.repositories_directory
     repository_directory = os.path.join(repositories_directory, self.repository_name)
-    current_path = os.path.join(repository_directory, 'metadata', 'current')
 
-    # Ensure the current path is valid/exists before saving it.
-    if not os.path.exists(current_path):
-      raise tuf.exceptions.RepositoryError('Missing'
-        ' ' + repr(current_path) + '.  This path must exist and, at a minimum,'
-        ' contain the Root metadata file.')
+    self.metadata_handler = metadata_handler_cls(repository_directory,
+                                                 self.repository_name, repository_mirrors)
 
-    self.metadata_directory['current'] = current_path
-
-    # Set the path for the previous set of metadata files.
-    previous_path = os.path.join(repository_directory, 'metadata', 'previous')
-
-    # Ensure the previous path is valid/exists.
-    if not os.path.exists(previous_path):
-      raise tuf.exceptions.RepositoryError('Missing ' + repr(previous_path) + '.'
-        '  This path MUST exist.')
-
-    self.metadata_directory['previous'] = previous_path
-
+    self.targets_handler = targets_handler_cls(repository_mirrors,
+                                               self.consistent_snapshot)
+  
     # Load current and previous metadata.
     for metadata_set in ['current', 'previous']:
       for metadata_role in ['root', 'targets', 'snapshot', 'timestamp']:
-        self._load_metadata_from_file(metadata_set, metadata_role)
+        self._load_metadata(metadata_set, metadata_role)
 
     # Raise an exception if the repository is missing the required 'root'
     # metadata.
@@ -782,7 +770,7 @@ class Updater(object):
 
 
 
-  def _load_metadata_from_file(self, metadata_set, metadata_role):
+  def _load_metadata(self, metadata_set, metadata_role):
     """
     <Purpose>
       Non-public method that loads current or previous metadata if there is a
@@ -820,34 +808,12 @@ class Updater(object):
 
     # Ensure we have a valid metadata set.
     if metadata_set not in ['current', 'previous']:
-      raise securesystemslib.exceptions.Error(
-          'Invalid metadata set: ' + repr(metadata_set))
+        raise securesystemslib.exceptions.Error(
+            'Invalid metadata set: ' + repr(metadata_set))
 
-    # Save and construct the full metadata path.
-    metadata_directory = self.metadata_directory[metadata_set]
-    metadata_filename = metadata_role + '.json'
-    metadata_filepath = os.path.join(metadata_directory, metadata_filename)
+    metadata_object = self.metadata_handler.load_metadata_object(metadata_set, metadata_role)
 
-    # Ensure the metadata path is valid/exists, else ignore the call.
-    if os.path.exists(metadata_filepath):
-      # Load the file.  The loaded object should conform to
-      # 'tuf.formats.SIGNABLE_SCHEMA'.
-      try:
-        metadata_signable = securesystemslib.util.load_json_file(
-            metadata_filepath)
-
-      # Although the metadata file may exist locally, it may not
-      # be a valid json file.  On the next refresh cycle, it will be
-      # updated as required.  If Root if cannot be loaded from disk
-      # successfully, an exception should be raised by the caller.
-      except securesystemslib.exceptions.Error:
-        return
-
-      tuf.formats.check_signable_object_format(metadata_signable)
-
-      # Extract the 'signed' role object from 'metadata_signable'.
-      metadata_object = metadata_signable['signed']
-
+    if metadata_object is not None:
       # Save the metadata object to the metadata store.
       self.metadata[metadata_set][metadata_role] = metadata_object
 
@@ -858,7 +824,6 @@ class Updater(object):
         if metadata_role == 'root':
           self._rebuild_key_and_role_db()
           self.consistent_snapshot = metadata_object['consistent_snapshot']
-
         elif metadata_object['_type'] == 'targets':
           # TODO: Should we also remove the keys of the delegated roles?
           self._import_delegations(metadata_role)
@@ -1053,7 +1018,6 @@ class Updater(object):
     # Raise 'securesystemslib.exceptions.FormatError' if the check fail.
     securesystemslib.formats.BOOLEAN_SCHEMA.check_match(
         unsafely_update_root_if_necessary)
-
     # Update the top-level metadata.  The _update_metadata_if_changed() and
     # _update_metadata() calls below do NOT perform an update if there
     # is insufficient trusted signatures for the specified metadata.
@@ -1061,7 +1025,7 @@ class Updater(object):
     root_metadata = self.metadata['current']['root']
 
     try:
-      self._ensure_not_expired(root_metadata, 'root')
+      self.metadata_handler._ensure_not_expired(root_metadata, 'root')
 
     except tuf.exceptions.ExpiredMetadataError:
       # Raise 'tuf.exceptions.NoWorkingMirrorError' if a valid (not
@@ -1129,13 +1093,9 @@ class Updater(object):
     <Returns>
       None.
     """
-
     # Retrieve the latest, remote root.json.
-    latest_root_metadata_file = self._get_metadata_file(
+    latest_root_metadata = self._get_metadata_file_content(
         'root', 'root.json', DEFAULT_ROOT_UPPERLENGTH, None)
-
-    latest_root_metadata = securesystemslib.util.load_json_string(
-        latest_root_metadata_file.read().decode('utf-8'))
 
 
     next_version = current_root_metadata['version'] + 1
@@ -1155,520 +1115,19 @@ class Updater(object):
 
 
 
-  def _check_hashes(self, file_object, trusted_hashes):
+  def _get_metadata_file_content(self, metadata_role, remote_filename,
+                         upperbound_filelength, expected_version,
+                         metadata_set='current'):
     """
-    <Purpose>
-      Non-public method that verifies multiple secure hashes of the downloaded
-      file 'file_object'.  If any of these fail it raises an exception.  This is
-      to conform with the TUF spec, which support clients with different hashing
-      algorithms. The 'hash.py' module is used to compute the hashes of
-      'file_object'.
-
-    <Arguments>
-      file_object:
-        A 'securesystemslib.util.TempFile' file-like object.  'file_object'
-        ensures that a read() without a size argument properly reads the entire
-        file.
-
-      trusted_hashes:
-        A dictionary with hash-algorithm names as keys and hashes as dict values.
-        The hashes should be in the hexdigest format.  Should be Conformant to
-        'securesystemslib.formats.HASHDICT_SCHEMA'.
-
-    <Exceptions>
-      securesystemslib.exceptions.BadHashError, if the hashes don't match.
-
-    <Side Effects>
-      Hash digest object is created using the 'securesystemslib.hash' module.
-
-    <Returns>
-      None.
+    Non-public method which returns content of a metadata file.
+    The file might not exist on disk. In case of using a bare git repository,
+    it is more efficient to just return the content of the file, than
+    to save that content to a file just so that it can later be read.
+    The specifics are handled by the metadata handler.
     """
-
-    # Verify each trusted hash of 'trusted_hashes'.  If all are valid, simply
-    # return.
-    for algorithm, trusted_hash in six.iteritems(trusted_hashes):
-      digest_object = securesystemslib.hash.digest(algorithm)
-      digest_object.update(file_object.read())
-      computed_hash = digest_object.hexdigest()
-
-      # Raise an exception if any of the hashes are incorrect.
-      if trusted_hash != computed_hash:
-        raise securesystemslib.exceptions.BadHashError(trusted_hash,
-            computed_hash)
-
-      else:
-        logger.info('The file\'s ' + algorithm + ' hash is'
-            ' correct: ' + trusted_hash)
-
-
-
-
-
-  def _hard_check_file_length(self, file_object, trusted_file_length):
-    """
-    <Purpose>
-      Non-public method that ensures the length of 'file_object' is strictly
-      equal to 'trusted_file_length'.  This is a deliberately redundant
-      implementation designed to complement
-      tuf.download._check_downloaded_length().
-
-    <Arguments>
-      file_object:
-        A 'securesystemslib.util.TempFile' file-like object.  'file_object'
-        ensures that a read() without a size argument properly reads the entire
-        file.
-
-      trusted_file_length:
-        A non-negative integer that is the trusted length of the file.
-
-    <Exceptions>
-      tuf.exceptions.DownloadLengthMismatchError, if the lengths do not match.
-
-    <Side Effects>
-      Reads the contents of 'file_object' and logs a message if 'file_object'
-      matches the trusted length.
-
-    <Returns>
-      None.
-    """
-
-    # Read the entire contents of 'file_object', a
-    # 'securesystemslib.util.TempFile' file-like object that ensures the entire
-    # file is read.
-    observed_length = len(file_object.read())
-
-    # Return and log a message if the length 'file_object' is equal to
-    # 'trusted_file_length', otherwise raise an exception.  A hard check
-    # ensures that a downloaded file strictly matches a known, or trusted,
-    # file length.
-    if observed_length != trusted_file_length:
-      raise tuf.exceptions.DownloadLengthMismatchError(trusted_file_length,
-          observed_length)
-
-    else:
-      logger.debug('Observed length (' + str(observed_length) +\
-          ') == trusted length (' + str(trusted_file_length) + ')')
-
-
-
-
-
-  def _soft_check_file_length(self, file_object, trusted_file_length):
-    """
-    <Purpose>
-      Non-public method that checks the trusted file length of a
-      'securesystemslib.util.TempFile' file-like object. The length of the file
-      must be less than or equal to the expected length. This is a deliberately
-      redundant implementation designed to complement
-      tuf.download._check_downloaded_length().
-
-    <Arguments>
-      file_object:
-        A 'securesystemslib.util.TempFile' file-like object.  'file_object'
-        ensures that a read() without a size argument properly reads the entire
-        file.
-
-      trusted_file_length:
-        A non-negative integer that is the trusted length of the file.
-
-    <Exceptions>
-      tuf.exceptions.DownloadLengthMismatchError, if the lengths do
-      not match.
-
-    <Side Effects>
-      Reads the contents of 'file_object' and logs a message if 'file_object'
-      is less than or equal to the trusted length.
-
-    <Returns>
-      None.
-    """
-
-    # Read the entire contents of 'file_object', a
-    # 'securesystemslib.util.TempFile' file-like object that ensures the entire
-    # file is read.
-    observed_length = len(file_object.read())
-
-    # Return and log a message if 'file_object' is less than or equal to
-    # 'trusted_file_length', otherwise raise an exception.  A soft check
-    # ensures that an upper bound restricts how large a file is downloaded.
-    if observed_length > trusted_file_length:
-      raise tuf.exceptions.DownloadLengthMismatchError(trusted_file_length,
-          observed_length)
-
-    else:
-      logger.debug('Observed length (' + str(observed_length) +\
-          ') <= trusted length (' + str(trusted_file_length) + ')')
-
-
-
-
-
-  def _get_target_file(self, target_filepath, file_length, file_hashes):
-    """
-    <Purpose>
-      Non-public method that safely (i.e., the file length and hash are
-      strictly equal to the trusted) downloads a target file up to a certain
-      length, and checks its hashes thereafter.
-
-    <Arguments>
-      target_filepath:
-        The target filepath (relative to the repository targets directory)
-        obtained from TUF targets metadata.
-
-      file_length:
-        The expected compressed length of the target file. If the file is not
-        compressed, then it will simply be its uncompressed length.
-
-      file_hashes:
-        The expected hashes of the target file.
-
-    <Exceptions>
-      tuf.exceptions.NoWorkingMirrorError:
-        The target could not be fetched. This is raised only when all known
-        mirrors failed to provide a valid copy of the desired target file.
-
-    <Side Effects>
-      The target file is downloaded from all known repository mirrors in the
-      worst case. If a valid copy of the target file is found, it is stored in
-      a temporary file and returned.
-
-    <Returns>
-      A 'securesystemslib.util.TempFile' file-like object containing the target.
-    """
-
-    # Define a callable function that is passed as an argument to _get_file()
-    # and called.  The 'verify_target_file' function ensures the file length
-    # and hashes of 'target_filepath' are strictly equal to the trusted values.
-    def verify_target_file(target_file_object):
-
-      # Every target file must have its length and hashes inspected.
-      self._hard_check_file_length(target_file_object, file_length)
-      self._check_hashes(target_file_object, file_hashes)
-
-    if self.consistent_snapshot:
-      # Note: values() does not return a list in Python 3.  Use list()
-      # on values() for Python 2+3 compatibility.
-      target_digest = list(file_hashes.values()).pop()
-      dirname, basename = os.path.split(target_filepath)
-      target_filepath = os.path.join(dirname, target_digest + '.' + basename)
-
-    return self._get_file(target_filepath, verify_target_file,
-        'target', file_length, download_safely=True)
-
-
-
-
-
-  def _verify_uncompressed_metadata_file(self, metadata_file_object,
-      metadata_role):
-    """
-    <Purpose>
-      Non-public method that verifies an uncompressed metadata file.  An
-      exception is raised if 'metadata_file_object is invalid.  There is no
-      return value.
-
-    <Arguments>
-      metadata_file_object:
-        A 'securesystemslib.util.TempFile' instance containing the metadata
-        file.  'metadata_file_object' ensures the entire file is returned with
-        read().
-
-      metadata_role:
-        The role name of the metadata (e.g., 'root', 'targets',
-        'unclaimed').
-
-    <Exceptions>
-      securesystemslib.exceptions.FormatError:
-        In case the metadata file is valid JSON, but not valid TUF metadata.
-
-      tuf.exceptions.InvalidMetadataJSONError:
-        In case the metadata file is not valid JSON.
-
-      tuf.exceptions.ReplayedMetadataError:
-        In case the downloaded metadata file is older than the current one.
-
-      tuf.exceptions.RepositoryError:
-        In case the repository is somehow inconsistent; e.g. a parent has not
-        delegated to a child (contrary to expectations).
-
-      tuf.SignatureError:
-        In case the metadata file does not have a valid signature.
-
-    <Side Effects>
-      The content of 'metadata_file_object' is read and loaded.
-
-    <Returns>
-      None.
-    """
-
-    metadata = metadata_file_object.read().decode('utf-8')
-
-    try:
-      metadata_signable = securesystemslib.util.load_json_string(metadata)
-
-    except Exception as exception:
-      raise tuf.exceptions.InvalidMetadataJSONError(exception)
-
-    else:
-      # Ensure the loaded 'metadata_signable' is properly formatted.  Raise
-      # 'securesystemslib.exceptions.FormatError' if not.
-      tuf.formats.check_signable_object_format(metadata_signable)
-
-    # Is 'metadata_signable' expired?
-    self._ensure_not_expired(metadata_signable['signed'], metadata_role)
-
-    # We previously verified version numbers in this function, but have since
-    # moved version number verification to the functions that retrieve
-    # metadata.
-
-    # Verify the signature on the downloaded metadata object.
-    valid = tuf.sig.verify(metadata_signable, metadata_role,
-        self.repository_name)
-
-    if not valid:
-      raise securesystemslib.exceptions.BadSignatureError(metadata_role)
-
-
-
-
-
-  def _get_metadata_file(self, metadata_role, remote_filename,
-    upperbound_filelength, expected_version):
-    """
-    <Purpose>
-      Non-public method that tries downloading, up to a certain length, a
-      metadata file from a list of known mirrors. As soon as the first valid
-      copy of the file is found, the downloaded file is returned and the
-      remaining mirrors are skipped.
-
-    <Arguments>
-      metadata_role:
-        The role name of the metadata (e.g., 'root', 'targets', 'unclaimed').
-
-      remote_filename:
-        The relative file path (on the remove repository) of 'metadata_role'.
-
-      upperbound_filelength:
-        The expected length, or upper bound, of the metadata file to be
-        downloaded.
-
-      expected_version:
-        The expected and required version number of the 'metadata_role' file
-        downloaded.  'expected_version' is an integer.
-
-    <Exceptions>
-      tuf.exceptions.NoWorkingMirrorError:
-        The metadata could not be fetched. This is raised only when all known
-        mirrors failed to provide a valid copy of the desired metadata file.
-
-    <Side Effects>
-      The file is downloaded from all known repository mirrors in the worst
-      case. If a valid copy of the file is found, it is stored in a temporary
-      file and returned.
-
-    <Returns>
-      A 'securesystemslib.util.TempFile' file-like object containing the
-      metadata.
-    """
-
-    file_mirrors = tuf.mirrors.get_list_of_mirrors('meta', remote_filename,
-        self.mirrors)
-
-    # file_mirror (URL): error (Exception)
-    file_mirror_errors = {}
-    file_object = None
-
-    for file_mirror in file_mirrors:
-      try:
-        file_object = tuf.download.unsafe_download(file_mirror,
-            upperbound_filelength)
-
-        # Verify 'file_object' according to the callable function.
-        # 'file_object' is also verified if decompressed above (i.e., the
-        # uncompressed version).
-        metadata_signable = \
-          securesystemslib.util.load_json_string(file_object.read().decode('utf-8'))
-
-        # Determine if the specification version number is supported.  It is
-        # assumed that "spec_version" is in (major.minor.fix) format, (for
-        # example: "1.4.3") and that releases with the same major version
-        # number maintain backwards compatibility.  Consequently, if the major
-        # version number of new metadata equals our expected major version
-        # number, the new metadata is safe to parse.
-        try:
-          spec_version_parsed = metadata_signable['signed']['spec_version'].split('.')
-          if int(spec_version_parsed[0]) != SUPPORTED_MAJOR_VERSION:
-            raise securesystemslib.exceptions.BadVersionNumberError('Downloaded'
-              ' metadata that specifies an unsupported spec_version.  Supported'
-              ' major version number: ' + repr(SUPPORTED_MAJOR_VERSION))
-
-        except (ValueError, TypeError):
-          raise securesystemslib.exceptions.FormatError('Improperly'
-            ' formatted spec_version, which must be in major.minor.fix format')
-
-        # If the version number is unspecified, ensure that the version number
-        # downloaded is greater than the currently trusted version number for
-        # 'metadata_role'.
-        version_downloaded = metadata_signable['signed']['version']
-
-        if expected_version is not None:
-          # Verify that the downloaded version matches the version expected by
-          # the caller.
-          if version_downloaded != expected_version:
-            raise securesystemslib.exceptions.BadVersionNumberError('Downloaded'
-              ' version number: ' + repr(version_downloaded) + '.  Version'
-              ' number MUST be: ' + repr(expected_version))
-
-        # The caller does not know which version to download.  Verify that the
-        # downloaded version is at least greater than the one locally
-        # available.
-        else:
-          # Verify that the version number of the locally stored
-          # 'timestamp.json', if available, is less than what was downloaded.
-          # Otherwise, accept the new timestamp with version number
-          # 'version_downloaded'.
-
-          try:
-            current_version = \
-              self.metadata['current'][metadata_role]['version']
-
-            if version_downloaded < current_version:
-              raise tuf.exceptions.ReplayedMetadataError(metadata_role,
-                  version_downloaded, current_version)
-
-          except KeyError:
-            logger.info(metadata_role + ' not available locally.')
-
-        self._verify_uncompressed_metadata_file(file_object, metadata_role)
-
-      except Exception as exception:
-        # Remember the error from this mirror, and "reset" the target file.
-        logger.exception('Update failed from ' + file_mirror + '.')
-        file_mirror_errors[file_mirror] = exception
-        file_object = None
-
-      else:
-        break
-
-    if file_object:
-      return file_object
-
-    else:
-      logger.error('Failed to update ' + repr(remote_filename) + ' from all'
-        ' mirrors: ' + repr(file_mirror_errors))
-      raise tuf.exceptions.NoWorkingMirrorError(file_mirror_errors)
-
-
-
-  def _verify_root_chain_link(self, rolename, current_root_metadata,
-    next_root_metadata):
-
-    if rolename != 'root':
-      return True
-
-    current_root_role = current_root_metadata['roles'][rolename]
-
-    # Verify next metadata with current keys/threshold
-    valid = tuf.sig.verify(next_root_metadata, rolename, self.repository_name,
-        current_root_role['threshold'], current_root_role['keyids'])
-
-    if not valid:
-      raise securesystemslib.exceptions.BadSignatureError('Root is not signed'
-          ' by previous threshold of keys.')
-
-
-
-
-
-  def _get_file(self, filepath, verify_file_function, file_type, file_length,
-      download_safely=True):
-    """
-    <Purpose>
-      Non-public method that tries downloading, up to a certain length, a
-      metadata or target file from a list of known mirrors. As soon as the first
-      valid copy of the file is found, the rest of the mirrors will be skipped.
-
-    <Arguments>
-      filepath:
-        The relative metadata or target filepath.
-
-      verify_file_function:
-        A callable function that expects a 'securesystemslib.util.TempFile'
-        file-like object and raises an exception if the file is invalid.
-        Target files and uncompressed versions of metadata may be verified with
-        'verify_file_function'.
-
-      file_type:
-        Type of data needed for download, must correspond to one of the strings
-        in the list ['meta', 'target'].  'meta' for metadata file type or
-        'target' for target file type.  It should correspond to the
-        'securesystemslib.formats.NAME_SCHEMA' format.
-
-      file_length:
-        The expected length, or upper bound, of the target or metadata file to
-        be downloaded.
-
-      download_safely:
-        A boolean switch to toggle safe or unsafe download of the file.
-
-    <Exceptions>
-      tuf.exceptions.NoWorkingMirrorError:
-        The metadata could not be fetched. This is raised only when all known
-        mirrors failed to provide a valid copy of the desired metadata file.
-
-    <Side Effects>
-      The file is downloaded from all known repository mirrors in the worst
-      case. If a valid copy of the file is found, it is stored in a temporary
-      file and returned.
-
-    <Returns>
-      A 'securesystemslib.util.TempFile' file-like object containing the
-      metadata or target.
-    """
-
-    file_mirrors = tuf.mirrors.get_list_of_mirrors(file_type, filepath,
-        self.mirrors)
-
-    # file_mirror (URL): error (Exception)
-    file_mirror_errors = {}
-    file_object = None
-
-    for file_mirror in file_mirrors:
-      try:
-        # TODO: Instead of the more fragile 'download_safely' switch, unroll
-        # the function into two separate ones: one for "safe" download, and the
-        # other one for "unsafe" download? This should induce safer and more
-        # readable code.
-        if download_safely:
-          file_object = tuf.download.safe_download(file_mirror, file_length)
-
-        else:
-          file_object = tuf.download.unsafe_download(file_mirror, file_length)
-
-        # Verify 'file_object' according to the callable function.
-        # 'file_object' is also verified if decompressed above (i.e., the
-        # uncompressed version).
-        verify_file_function(file_object)
-
-      except Exception as exception:
-        # Remember the error from this mirror, and "reset" the target file.
-        logger.exception('Update failed from ' + file_mirror + '.')
-        file_mirror_errors[file_mirror] = exception
-        file_object = None
-
-      else:
-        break
-
-    if file_object:
-      return file_object
-
-    else:
-      logger.error('Failed to update ' + repr(filepath) + ' from'
-          ' all mirrors: ' + repr(file_mirror_errors))
-      raise tuf.exceptions.NoWorkingMirrorError(file_mirror_errors)
-
-
+    current_version = self.metadata[metadata_set][metadata_role]['version']
+    return self.metadata_handler.get_metadata_file_content(
+        metadata_role, metadata_role + '.json', None, current_version, DEFAULT_ROOT_UPPERLENGTH)
 
 
 
@@ -1710,80 +1169,34 @@ class Updater(object):
     <Returns>
       None.
     """
+    current_version = None
+    try:
+      current_version = self.metadata['current'][metadata_role]['version']
+    except KeyError:
+      pass
 
-    # Construct the metadata filename as expected by the download/mirror
-    # modules.
     metadata_filename = metadata_role + '.json'
-    metadata_filename = metadata_filename
-
-    # Attempt a file download from each mirror until the file is downloaded and
-    # verified.  If the signature of the downloaded file is valid, proceed,
-    # otherwise log a warning and try the next mirror.  'metadata_file_object'
-    # is the file-like object returned by 'download.py'.  'metadata_signable'
-    # is the object extracted from 'metadata_file_object'.  Metadata saved to
-    # files are regarded as 'signable' objects, conformant to
-    # 'tuf.formats.SIGNABLE_SCHEMA'.
-    #
-    # Some metadata (presently timestamp) will be downloaded "unsafely", in the
-    # sense that we can only estimate its true length and know nothing about
-    # its version.  This is because not all metadata will have other metadata
-    # for it; otherwise we will have an infinite regress of metadata signing
-    # for each other. In this case, we will download the metadata up to the
-    # best length we can get for it, not request a specific version, but
-    # perform the rest of the checks (e.g., signature verification).
-
-    remote_filename = metadata_filename
-    filename_version = ''
-
-    if self.consistent_snapshot and version:
-      filename_version = version
-      dirname, basename = os.path.split(remote_filename)
-      remote_filename = os.path.join(
-          dirname, str(filename_version) + '.' + basename)
-
-    metadata_file_object = \
-      self._get_metadata_file(metadata_role, remote_filename,
-        upperbound_filelength, version)
-
-    # The metadata has been verified. Move the metadata file into place.
-    # First, move the 'current' metadata file to the 'previous' directory
-    # if it exists.
-    current_filepath = os.path.join(self.metadata_directory['current'],
-                metadata_filename)
-    current_filepath = os.path.abspath(current_filepath)
-    securesystemslib.util.ensure_parent_dir(current_filepath)
-
-    previous_filepath = os.path.join(self.metadata_directory['previous'],
-        metadata_filename)
-    previous_filepath = os.path.abspath(previous_filepath)
-
-    if os.path.exists(current_filepath):
-      # Previous metadata might not exist, say when delegations are added.
-      securesystemslib.util.ensure_parent_dir(previous_filepath)
-      shutil.move(current_filepath, previous_filepath)
-
-    # Next, move the verified updated metadata file to the 'current' directory.
-    # Note that the 'move' method comes from securesystemslib.util's TempFile class.
-    # 'metadata_file_object' is an instance of securesystemslib.util.TempFile.
-    metadata_signable = \
-      securesystemslib.util.load_json_string(metadata_file_object.read().decode('utf-8'))
-
-    metadata_file_object.move(current_filepath)
+    metadata_signable = self.metadata_handler.get_updated_metadata(metadata_role,
+                                               current_version, version, self.consistent_snapshot,
+                                               upperbound_filelength)
+    updated_metadata_object = metadata_signable['signed']
 
     # Extract the metadata object so we can store it to the metadata store.
     # 'current_metadata_object' set to 'None' if there is not an object
     # stored for 'metadata_role'.
-    updated_metadata_object = metadata_signable['signed']
+
     current_metadata_object = self.metadata['current'].get(metadata_role)
 
-    self._verify_root_chain_link(metadata_role, current_metadata_object,
+    self.metadata_handler._verify_root_chain_link(metadata_role, current_metadata_object,
         metadata_signable)
 
     # Finally, update the metadata and fileinfo stores, and rebuild the
     # key and role info for the top-level roles if 'metadata_role' is root.
     # Rebuilding the the key and role info is required if the newly-installed
     # root metadata has revoked keys or updated any top-level role information.
-    logger.debug('Updated ' + repr(current_filepath) + '.')
+    
+    logger.debug('Updated ' + repr(metadata_filename) + '.')
+
     self.metadata['previous'][metadata_role] = current_metadata_object
     self.metadata['current'][metadata_role] = updated_metadata_object
     self._update_versioninfo(metadata_filename)
@@ -1850,7 +1263,6 @@ class Updater(object):
     <Returns>
       None.
     """
-
     metadata_filename = metadata_role + '.json'
     expected_versioninfo = None
 
@@ -1883,7 +1295,7 @@ class Updater(object):
       # check to see if our local version is stale and notify the user if so.
       # This raises tuf.exceptions.ExpiredMetadataError if the metadata we have
       # is expired. Resolves issue #322.
-      self._ensure_not_expired(self.metadata['current'][metadata_role],
+      self.metadata_handler._ensure_not_expired(self.metadata['current'][metadata_role],
           metadata_role)
 
       # TODO: If 'metadata_role' is root or snapshot, we should verify that
@@ -2026,17 +1438,12 @@ class Updater(object):
     <Returns>
       None.
     """
-
     # In case we delayed loading the metadata and didn't do it in
     # __init__ (such as with delegated metadata), then get the version
     # info now.
 
     # Save the path to the current metadata file for 'metadata_filename'.
-    current_filepath = os.path.join(self.metadata_directory['current'],
-        metadata_filename)
-
-    # If the path is invalid, simply return and leave versioninfo unset.
-    if not os.path.exists(current_filepath):
+    if not self.metadata_handler.metadata_file_exists('current', metadata_filename):
       self.versioninfo[metadata_filename] = None
       return
 
@@ -2069,6 +1476,7 @@ class Updater(object):
     else:
 
       try:
+
         # The metadata file names in 'self.metadata' exclude the role
         # extension.  Strip the '.json' extension when checking if
         # 'metadata_filename' currently exists.
@@ -2190,64 +1598,15 @@ class Updater(object):
     # info now.
 
     # Save the path to the current metadata file for 'metadata_filename'.
-    current_filepath = os.path.join(self.metadata_directory['current'],
-        metadata_filename)
-
-    # If the path is invalid, simply return and leave fileinfo unset.
-    if not os.path.exists(current_filepath):
+    if not self.metadata_handler.metadata_file_exists('current', metadata_filename):
       self.fileinfo[metadata_filename] = None
       return
 
     # Extract the file information from the actual file and save it
     # to the fileinfo store.
-    file_length, hashes = securesystemslib.util.get_file_details(
-        current_filepath)
+    file_length, hashes = self.metadata_handler.get_metadata_file_details('current', metadata_filename)
     metadata_fileinfo = tuf.formats.make_fileinfo(file_length, hashes)
     self.fileinfo[metadata_filename] = metadata_fileinfo
-
-
-
-
-
-
-
-  def _move_current_to_previous(self, metadata_role):
-    """
-    <Purpose>
-      Non-public method that moves the current metadata file for 'metadata_role'
-      to the previous directory.
-
-    <Arguments>
-      metadata_role:
-        The name of the metadata. This is a role name and should not end
-        in '.json'.  Examples: 'root', 'targets', 'targets/linux/x86'.
-
-    <Exceptions>
-      None.
-
-    <Side Effects>
-     The metadata file for 'metadata_role' is removed from 'current'
-     and moved to the 'previous' directory.
-
-    <Returns>
-      None.
-    """
-
-    # Get the 'current' and 'previous' full file paths for 'metadata_role'
-    metadata_filepath = metadata_role + '.json'
-    previous_filepath = os.path.join(self.metadata_directory['previous'],
-                                     metadata_filepath)
-    current_filepath = os.path.join(self.metadata_directory['current'],
-                                    metadata_filepath)
-
-    # Remove the previous path if it exists.
-    if os.path.exists(previous_filepath):
-      os.remove(previous_filepath)
-
-    # Move the current path to the previous path.
-    if os.path.exists(current_filepath):
-      securesystemslib.util.ensure_parent_dir(previous_filepath)
-      os.rename(current_filepath, previous_filepath)
 
 
 
@@ -2279,70 +1638,14 @@ class Updater(object):
 
     # The root metadata role is never deleted without a replacement.
     if metadata_role == 'root':
-      return
+        return
 
-    # Get rid of the current metadata file.
-    self._move_current_to_previous(metadata_role)
-
+    self.metadata_handler.delete_metadata(metadata_role)
     # Remove knowledge of the role.
     if metadata_role in self.metadata['current']:
       del self.metadata['current'][metadata_role]
+
     tuf.roledb.remove_role(metadata_role, self.repository_name)
-
-
-
-
-
-  def _ensure_not_expired(self, metadata_object, metadata_rolename):
-    """
-    <Purpose>
-      Non-public method that raises an exception if the current specified
-      metadata has expired.
-
-    <Arguments>
-      metadata_object:
-        The metadata that should be expired, a 'tuf.formats.ANYROLE_SCHEMA'
-        object.
-
-      metadata_rolename:
-        The name of the metadata. This is a role name and should not end
-        in '.json'.  Examples: 'root', 'targets', 'targets/linux/x86'.
-
-    <Exceptions>
-      tuf.exceptions.ExpiredMetadataError:
-        If 'metadata_rolename' has expired.
-
-    <Side Effects>
-      None.
-
-    <Returns>
-      None.
-    """
-
-    # Extract the expiration time.
-    expires = metadata_object['expires']
-
-    # If the current time has surpassed the expiration date, raise an
-    # exception.  'expires' is in
-    # 'securesystemslib.formats.ISO8601_DATETIME_SCHEMA' format (e.g.,
-    # '1985-10-21T01:22:00Z'.)  Convert it to a unix timestamp and compare it
-    # against the current time.time() (also in Unix/POSIX time format, although
-    # with microseconds attached.)
-    current_time = int(time.time())
-
-    # Generate a user-friendly error message if 'expires' is less than the
-    # current time (i.e., a local time.)
-    expires_datetime = iso8601.parse_date(expires)
-    expires_timestamp = tuf.formats.datetime_to_unix_timestamp(expires_datetime)
-
-    if expires_timestamp < current_time:
-      message = 'Metadata '+repr(metadata_rolename)+' expired on ' + \
-        expires_datetime.ctime() + ' (UTC).'
-      logger.error(message)
-
-      raise tuf.exceptions.ExpiredMetadataError(message)
-
-
 
 
 
@@ -2464,7 +1767,6 @@ class Updater(object):
 
         else:
           continue
-
     # If there is nothing to refresh, we are done.
     if not roles_to_update:
       return
@@ -2474,8 +1776,8 @@ class Updater(object):
     # Iterate 'roles_to_update', and load and update its metadata file if it
     # has changed.
     for rolename in roles_to_update:
-      self._load_metadata_from_file('previous', rolename)
-      self._load_metadata_from_file('current', rolename)
+      self._load_metadata('previous', rolename)
+      self._load_metadata('current', rolename)
 
       self._update_metadata_if_changed(rolename)
 
@@ -2520,7 +1822,6 @@ class Updater(object):
       targets of 'rolename'.  Conformant to
       'tuf.formats.TARGETINFOS_SCHEMA'.
     """
-
     if targets is None:
       targets = []
 
@@ -2643,7 +1944,6 @@ class Updater(object):
       The target information for 'target_filepath', conformant to
       'tuf.formats.TARGETINFO_SCHEMA'.
     """
-
     # Does 'target_filepath' have the correct format?
     # Raise 'securesystemslib.exceptions.FormatError' if there is a mismatch.
     securesystemslib.formats.RELPATH_SCHEMA.check_match(target_filepath)
@@ -2656,7 +1956,6 @@ class Updater(object):
 
     # Get target by looking at roles in order of priority tags.
     target = self._preorder_depth_first_walk(target_filepath)
-
     # Raise an exception if the target information could not be retrieved.
     if target is None:
       logger.error(repr(target_filepath) + ' not found.')
@@ -2749,6 +2048,7 @@ class Updater(object):
         child_roles_to_visit = []
         # NOTE: This may be a slow operation if there are many delegated roles.
         for child_role in child_roles:
+
           child_role_name = self._visit_child_role(child_role, target_filepath)
           if child_role['terminating'] and child_role_name is not None:
             logger.debug('Adding child role ' + repr(child_role_name))
@@ -2870,7 +2170,6 @@ class Updater(object):
 
       Otherwise, we return None.
     """
-
     child_role_name = child_role['name']
     child_role_paths = child_role.get('paths')
     child_role_path_hash_prefixes = child_role.get('path_hash_prefixes')
@@ -2961,8 +2260,6 @@ class Updater(object):
 
 
 
-
-
   def remove_obsolete_targets(self, destination_directory):
     """
     <Purpose>
@@ -3000,33 +2297,13 @@ class Updater(object):
           for target in self.metadata['previous'][role]['targets']:
             if target not in self.metadata['current'][role]['targets']:
               # 'target' is only in 'previous', so remove it.
-              logger.warning('Removing obsolete file: ' + repr(target) + '.')
-
-              # Remove the file if it hasn't been removed already.
-              destination = \
-                os.path.join(destination_directory, target.lstrip(os.sep))
-              try:
-                os.remove(destination)
-
-              except OSError as e:
-                # If 'filename' already removed, just log it.
-                if e.errno == errno.ENOENT:
-                  logger.info('File ' + repr(destination) + ' was already'
-                    ' removed.')
-
-                else:
-                  logger.error(str(e))
-
+  	            self.targets_handler._remove_obsolete_target(destination_directory, target)
             else:
               logger.debug('Skipping: ' + repr(target) + '.  It is still'
                 ' a current target.')
         else:
           logger.debug('Skipping: ' + repr(role) + '.  Not in the previous'
             ' metadata')
-
-
-
-
 
   def updated_targets(self, targets, destination_directory):
     """
@@ -3067,128 +2344,8 @@ class Updater(object):
     # Raise 'securesystemslib.exceptions.FormatError' if there is a mismatch.
     tuf.formats.TARGETINFOS_SCHEMA.check_match(targets)
     securesystemslib.formats.PATH_SCHEMA.check_match(destination_directory)
-
-    # Keep track of the target objects and filepaths of updated targets.
-    # Return 'updated_targets' and use 'updated_targetpaths' to avoid
-    # duplicates.
-    updated_targets = []
-    updated_targetpaths = []
-
-    for target in targets:
-      # Prepend 'destination_directory' to the target's relative filepath (as
-      # stored in metadata.)  Verify the hash of 'target_filepath' against
-      # each hash listed for its fileinfo.  Note: join() discards
-      # 'destination_directory' if 'filepath' contains a leading path separator
-      # (i.e., is treated as an absolute path).
-      filepath = target['filepath']
-      if filepath[0] == '/':
-        filepath = filepath[1:]
-      target_filepath = os.path.join(destination_directory, filepath)
-
-      if target_filepath in updated_targetpaths:
-        continue
-
-      # Try one of the algorithm/digest combos for a mismatch.  We break
-      # as soon as we find a mismatch.
-      for algorithm, digest in six.iteritems(target['fileinfo']['hashes']):
-        digest_object = None
-        try:
-          digest_object = securesystemslib.hash.digest_filename(target_filepath,
-            algorithm=algorithm)
-
-        # This exception would occur if the target does not exist locally.
-        except IOError:
-          updated_targets.append(target)
-          updated_targetpaths.append(target_filepath)
-          break
-
-        # The file does exist locally, check if its hash differs.
-        if digest_object.hexdigest() != digest:
-          updated_targets.append(target)
-          updated_targetpaths.append(target_filepath)
-          break
-
-    return updated_targets
-
-
-
-
+    return self.targets_handler._get_updated_targets(targets, destination_directory)
 
   def download_target(self, target, destination_directory):
-    """
-    <Purpose>
-      Download 'target' and verify it is trusted.
+    self.targets_handler.download_target(target, destination_directory)
 
-      This will only store the file at 'destination_directory' if the
-      downloaded file matches the description of the file in the trusted
-      metadata.
-
-    <Arguments>
-      target:
-        The target to be downloaded.  Conformant to
-        'tuf.formats.TARGETINFO_SCHEMA'.
-
-      destination_directory:
-        The directory to save the downloaded target file.
-
-    <Exceptions>
-      securesystemslib.exceptions.FormatError:
-        If 'target' is not properly formatted.
-
-      tuf.exceptions.NoWorkingMirrorError:
-        If a target could not be downloaded from any of the mirrors.
-
-        Although expected to be rare, there might be OSError exceptions (except
-        errno.EEXIST) raised when creating the destination directory (if it
-        doesn't exist).
-
-    <Side Effects>
-      A target file is saved to the local system.
-
-    <Returns>
-      None.
-    """
-
-    # Do the arguments have the correct format?
-    # This check ensures the arguments have the appropriate
-    # number of objects and object types, and that all dict
-    # keys are properly named.
-    # Raise 'securesystemslib.exceptions.FormatError' if the check fail.
-    tuf.formats.TARGETINFO_SCHEMA.check_match(target)
-    securesystemslib.formats.PATH_SCHEMA.check_match(destination_directory)
-
-    # Extract the target file information.
-    target_filepath = target['filepath']
-    trusted_length = target['fileinfo']['length']
-    trusted_hashes = target['fileinfo']['hashes']
-
-    # '_get_target_file()' checks every mirror and returns the first target
-    # that passes verification.
-    target_file_object = self._get_target_file(target_filepath, trusted_length,
-        trusted_hashes)
-
-    # We acquired a target file object from a mirror.  Move the file into place
-    # (i.e., locally to 'destination_directory').  Note: join() discards
-    # 'destination_directory' if 'target_path' contains a leading path
-    # separator (i.e., is treated as an absolute path).
-    destination = os.path.join(destination_directory,
-        target_filepath.lstrip(os.sep))
-    destination = os.path.abspath(destination)
-    target_dirpath = os.path.dirname(destination)
-
-    # When attempting to create the leaf directory of 'target_dirpath', ignore
-    # any exceptions raised if the root directory already exists.  All other
-    # exceptions potentially thrown by os.makedirs() are re-raised.
-    # Note: os.makedirs can raise OSError if the leaf directory already exists
-    # or cannot be created.
-    try:
-      os.makedirs(target_dirpath)
-
-    except OSError as e:
-      if e.errno == errno.EEXIST:
-        pass
-
-      else:
-        raise
-
-    target_file_object.move(destination)
